mbe: Support `vis` fragment field for `fn`; add infra for recursive token collection
Add support for getting a `.vis` field from a `fn` type. This requires
adding support for collecting tokens when parsing the visibility of a
function. Add a new `ForceCollect::Recursive`, and plumb that through
the item parser enough to support visibility. Only use that when a macro
uses macro metavariable expressions, as we don't need the tokens
otherwise.

`ForceCollect::Recursive` will need further handling later, to capture
the tokens of any other fields needed, but that can be added alongside
any fields it supports.