rustc_target: RISC-V: Add two supervisor extensions for intrinsics
This commit adds two ratified **privileged** RISC-V extensions
corresponding to 35 existing intrinsics in `core::arch::riscv{32,64}`.

Because they semantically require the H and/or Svinval privileged
extensions, we'd better to have those new target features to use them
as a part of the `#[target_feature(enable = ...)]` attribute.

It also adds a note about a conflict between `e` and `h` (no code
changes for now but should be rejected in the future).