Rename `range_search` to `core::unicode::rt`
Instead of `include_str!()`ing `range_search.rs`, just make it a normal
module under `core::unicode`. This means the same source code doesn't
have to be checked in twice, and it plays nicer with IDEs.

Also rename it to `rt` because it will also include runtime functions
for case foldings in the next commit.