x86: Correctly pass large integers in registers with -Zregparm
The -Zregparm option for 32-bit x86 modifies the calling convention to pass
a number of arguments in registers instead of on the stack. (This is primarily
used by the Linux kernel.)

Currently, rustc will only pass integers of size <= 32 bits in registers, but
gcc (and clang) will pass larger integers (e.g., 64-bit integers) across two
registers if possible. This is not the case with fastcall/vectorcall.

Note that gcc/clang with -mregparm will also pass some structs in registers:
this patch does not attempt to fix that case.