x86: Correctly pass large integers and structs in registers with -Zregparm
The -Zregparm option for 32-bit x86 modifies the calling convention to pass
a number of arguments in registers instead of on the stack. (This is primarily
used by the Linux kernel.)

Currently, rustc will only pass integers of size <= 32 bits in registers, but
gcc (and clang) will pass larger integers (e.g., 64-bit integers) across two
registers if possible. This is not the case with fastcall/vectorcall.

Equally, struct arguments will be split into registers and passed where
possible.

Supporting this in rustc required moving the 'inreg' determination into
compute_abi_info, as this is the stage where we determine if structs are
passed directly or indirectly. This involves updating the x86_win32 ABI
implementation as well, as it previously shared the inreg handling, but has
a different compute_abi_info implementation.

Also add some assembly-llvm and codegen-llvm tests. Note that the LLVM
IR generated doesn't exactly match clang, as rustc treats struct arguments
as one argument in three registers, whereas clang seems to decompose it
into three registers. It seems to work regardless...