Allow `build-fail` tests to optionally skip the automatic `check-pass` step
By default, all `//@ build-fail` UI tests must first succeed in `check-pass`
mode, because build-fail normally only makes sense for tests that would check
successfully.

However, in some cases a full build will fail in a different way, e.g. due to
an ICE that is not present in check builds. For those tests, it's useful to be
able to opt out of the check-pass step.