rustdoc: add support for macro_rules macros of multiple kinds
This adds support for advanced RFC 3697 and 3698 macros, that look
like this:

```
/// Doc Comment.
macro_rules! NAME {
    attr(key = $value:literal) ($attached:item) => { ... };
    derive() ($attached:item) => { ... };
    ($bang:tt) => { ... };
}
```

And can be used like this:

```
/*attr*/ #[NAME]
/*derive*/ #[derive(NAME)]
/*bang*/ NAME!{}
```

The basic problem is that there are three different ways to use
this macro, and, historically, these three ways were implemented
by fully-separated proc macro items with separate doc comments.
We want these new declarative macros to work the same way,
so they appear in the same section of the module table of contents
and can be filtered for in the search engine.

This commit makes the feature work by generating separate pages
for each macro kind. These separate pages appear in separate
sections of the module table of contents, produce separate
search results, and have duplicate copies of the doc comment.

The following features would also be helpful, but are left
for future commits to implement:

- To improve the produced documentation, we should probably add
  support for adding `///` attributes to macro rules arms, so that
  the different syntaxes can actually be documented separately.

- This commit doesn't test intra-doc links.

Parts of https://github.com/rust-lang/rust/pull/145458 are directly
used in this code.

Co-Authored-By: Guillaume Gomez <guillaume1.gomez@gmail.com>