rustc_codegen_llvm: Add debug locations to `write_operand_repeatedly()`
So that stepping with a debugger over lines such as:

    let h = ["whatever"; 8];

works.

Use the same technique as we already do in `cfi_type_test()`.

Since all emitted LLVM instructions are looped over, a debugger will
unfortunately step as many times as we repeat. Not sure how to solve
that. It can become quite annoying if you have say `1024` repetitions.