Ensure the personality does not panic
In a cdylib that uses std and is free from panics in the code, the panic
machinery will *still* be pulled in because of the personality function
when using LLD.

The personality function is used by unwinding to figure out what to do
when unwinding through a function. Each function that participates in
unwind has an associated FDE (frame descriptor entries) in `.eh_frame`.
This FDE points to a CIE (common information entry), which can reference
a language-specific personality function, like `rust_eh_personality` in
our case.

As long as there is a CIE that references the personality, the function
cannot be removed. If all references to a CIE get removed (because the
functions and their associated FDEs have been removed), LLD will not
remove the CIE, likely due to the ordering of its passes). Binutils ld
will remove it.

In the case where the CIE is still around (despite not being used), it
will still reference the personality function, so that will still be
around. This is not great since it's a bunch of code, but also not
_that_ much. But this is where panicking comes in.

Before this change, the personality function internally made use of `dyn
Fn`. This caused an indirect call that LLVM was not able to analyze as
guaranteed free of unwinding, even during fat LTO. This meant that an
`invoke` was used, with a landing pad. In an `extern "C"` function,
which the personality function is, all landing pads call
`panic_cannot_unwind`, which is a `panic_nounwind`, which is, obviously,
a panic. And as a panic, it pulls in *all* the panic machinery, which is
very big and sad.

It is also completely unnecessary, because these indirect functions do
not panic, as they are just a convenient abstraction provided from the
outside. By restructuring the code to remove these indirect calls, LLVM
is able to fully analyze everything and see that rust_eh_personality
cannot panic, and therefore remove its landing pad.

With this change, exporting a panic-free function from a cdylib will
only contain the function and the personality (when linked with LLD at
least, with binutils ld it will only contain the function), with no
panic code being present at all, which is great.