library: core: document layout guarantee of `TypeId`
Document the layout guarantee of `TypeId` not to exceed 16 bytes.
Additionally, use `repr(C)` to avoid additional padding if layout
randomization improved.

This is useful for FFI use-cases where a `TypeId` may be stored within a
foreign data structure or buffer.

An example for such a use-case can be found in [1].

Link: https://git.kernel.org/pub/scm/linux/kernel/git/driver-core/driver-core.git/commit/?id=6f61a2637abe4f89877da3280775565baedb60e0 [1]
Suggested-by: Alice Ryhl <aliceryhl@google.com>
Signed-off-by: Danilo Krummrich <dakr@kernel.org>